+++ packages/browser-wallet/src/popup/pages/ViewSeedPhrase/ViewSeedPhrase.tsx
@@ -4,7 +4,7 @@ import Form from '@popup/shared/Form';
 import FormPassword from '@popup/shared/Form/Password';
 import Submit from '@popup/shared/Form/Submit';
 import { TextArea } from '@popup/shared/Form/TextArea';
-import { encryptedSeedPhraseAtom, sessionPasscodeAtom } from '@popup/store/settings';
+import { encryptedSeedPhraseAtom } from '@popup/store/settings';
 import { useAtomValue } from 'jotai';
 import React, { useState } from 'react';
 import { Validate } from 'react-hook-form';
@@ -13,13 +13,14 @@ import { useNavigate } from 'react-router-dom';
 import { absoluteRoutes } from '@popup/constants/routes';
 import { useAsyncMemo } from 'wallet-common-helpers';
 import { decrypt } from '@shared/utils/crypto';
+import { MemoryStoreKeys, useMemoryStoreValue } from '@popup/shared/utils/background-storage-helpers';
 
 export default function ViewSeedPhrase() {
     const nav = useNavigate();
     const { t: tSetup } = useTranslation('setup');
     const { t: tPasscode } = useTranslation('changePasscode');
     const { t } = useTranslation('viewSeedPhrase');
-    const passcode = useAtomValue(sessionPasscodeAtom);
+    const passcode = useMemoryStoreValue(MemoryStoreKeys.Passcode);
     const [showSeedPhrase, setShowSeedPhrase] = useState<boolean>(false);
     const encryptedSeed = useAtomValue(encryptedSeedPhraseAtom);